BattleScene_Update
_BattleScene
-Healsystem entfernt und wird sich nur noch auf den CharacterStatus konzentriert
- HP Balken funktionieren mit nicer animation
- Damage wird acordingly wiedergeben wenn Schaden entsteht
- aniamtion und eventhandler hinzugefügt

+++ Assets/_Game/Script/DamageScript.cs
@@ -6,7 +6,7 @@ public class DamageScript : MonoBehaviour
 {
 
     public StatusManager status;
-    public int damage = 50;
+    public float damage = 50;
     // Start is called before the first frame update
